requirements.txt
PyQt6==6.11.0
PyInstaller==6.19.0
playwright==1.58.0
screeninfo==0.8.1
requests==2.32.3
PySocks==1.7.1
Nuitka==2.8.3
Pillow
winshell; platform_system=="Windows"
pywin32; platform_system=="Windows"
